+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Unreleased]
+## [0.6.0] - 2023-07-25
 
 ### Added
 
@@ -70,7 +70,8 @@
 
 - Changed `EntryNotFoundException` parent from `\InvalidArgumentException` to `NotFoundException`
 
-[Unreleased]: https://github.com/phetit/dependency-injection/compare/v0.5.0...main
+[Unreleased]: https://github.com/phetit/dependency-injection/compare/v0.6.0...main
+[0.6.0]: https://github.com/phetit/dependency-injection/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/phetit/dependency-injection/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/phetit/dependency-injection/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/phetit/dependency-injection/compare/v0.2.0...v0.3.0